split EditHomeView into multiple files
