+++ source/views/home/edit/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export {ConnectedEditHomeView} from './list'
